requirements.txt
# Silver Tier Python dependencies
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
playwright>=1.40.0
watchdog>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
